% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.LeveneT2.R
\name{print.LeveneT2}
\alias{print.LeveneT2}
\title{Prints Levene's test based on Hotelling's \eqn{T^2} test}
\usage{
\method{print}{LeveneT2}(x, long = FALSE, ...)
}
\arguments{
\item{x}{an object of class \code{"LeveneT2"}}

\item{long}{a logical variable indicating whether a long output is desired
(\code{TRUE}) or not (\code{FALSE}, the default)}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Prints the results produced by \code{\link[smsets]{LeveneT2}}, consisting of
Levene's test for two multivariate samples based on Hotelling's \eqn{T^2}
test
}
\examples{
data(sparrows)
LeveneT2.sparrows <- LeveneT2(sparrows, group = Survivorship, level1 = "S",
                              var.equal = TRUE)
# Long output
print(LeveneT2.sparrows, long = TRUE)

}
