% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.Levenetests2s.mv.R
\name{print.Levenetests2s.mv}
\alias{print.Levenetests2s.mv}
\title{Prints multiple two-sample Levene tests for the comparison of variation in
multivariate data}
\usage{
\method{print}{Levenetests2s.mv}(x, ...)
}
\arguments{
\item{x}{an object of class "Levene.t"}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Prints the results produced by \code{\link[smsets]{Levenetests2s.mv}},
consisting of two-sample Levene's tests computed from two-sample t-tests
applied to absolute differences around medians for more than one response
vector
}
\details{
Summarize
}
\examples{
data(sparrows)
res.Levene2s.mv <- Levenetests2s.mv(sparrows, Survivorship, "S",
                               alternative = "less", var.equal = TRUE,
                               P.adjust = "bonferroni", unit = "mm")
print(res.Levene2s.mv)

}
