% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.OnewayMANOVA.R
\name{print.OnewayMANOVA}
\alias{print.OnewayMANOVA}
\title{Prints a one-way MANOVA with extra information}
\usage{
\method{print}{OnewayMANOVA}(
  x,
  test = c("Pillai", "Wilks", "Hotelling-Lawley", "Roy"),
  long = FALSE,
  ...
)
}
\arguments{
\item{x}{An object of class \code{OnewayMANOVA}.}

\item{test}{The name of the test statistic to be used (the four tests
implemented in \code{summary.manova}). Pillai's test is the default. Partial
matching is used so the name can be abbreviated.}

\item{long}{A logical variable indicating whether a long output is desired
(\code{TRUE}) or not (\code{FALSE}, the default)}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Prints the results produced by the \code{\link[smsets]{OnewayMANOVA}}
function
}
\examples{
data(skulls)
res.MANOVA <- OnewayMANOVA(skulls, group = Period)
# Long output, Wilks' test
print(res.MANOVA, test = "Wilks", long = TRUE)

}
