% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.Penrose.dist.R
\name{print.Penrose.dist}
\alias{print.Penrose.dist}
\title{Prints Penrose's distance matrix}
\usage{
\method{print}{Penrose.dist}(x, long = FALSE, ...)
}
\arguments{
\item{x}{an object of class Penrose.dist}

\item{long}{a logical variable indicating whether a long output is desired
(\code{TRUE}) or not (\code{FALSE}, the default). In addition to Penrose's
distances, the long output displays the covariance matrix for each group with
their sample sizes, the mean vector for each group, and the pooled covariance
matrix.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Prints the results produced by Penrose.dist
}
\examples{
data(skulls)
res.Penrose <- Penrose.dist(x = skulls, group = Period)
# Long output
print(res.Penrose, long = TRUE)

}
