% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.ttests2s.mv.R
\name{print.ttests2s.mv}
\alias{print.ttests2s.mv}
\title{Prints multiple two-sample t-tests for a multivariate data set}
\usage{
\method{print}{ttests2s.mv}(x, ...)
}
\arguments{
\item{x}{an object of class \code{"ttests"}}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Prints the results produced by \code{\link[smsets]{ttests2s.mv}}, consisting
of two-sample t-tests on more than one response vector with corrected
significance levels for multiple comparisons, as offered by \code{p.adjust}.
Effects sizes are also displayed.
}
\examples{
data(sparrows)
ttests.sparrows <- ttests2s.mv(sparrows, group = Survivorship, level1 = "S",
                              var.equal = TRUE, P.adjust = "holm",
                              unit = "mm")
print(ttests.sparrows)

}
