% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smsets-package.R
\docType{package}
\name{smsets-package}
\alias{smsets}
\alias{smsets-package}
\title{smsets}
\description{
A collection of simple parameter estimation and tests for the comparison of multivariate means and variation, to accompany Chapters 4 and 5 of the book Multivariate Statistical Methods. A Primer (5th edition), by Manly BFJ, Navarro Alberto JA & Gerow K (2024) \doi{10.1201/9781003453482}.
}
\details{
The goal of smsets is to produce simple multivariate statistical tests for
means and variances / covariances for one single factor with two or more
levels, including multiple two-sample t- and Levene’s tests, Hotelling’s
\eqn{T^2} test, extended two-sample and m-sample Levene’s tests for
multivariate data, one-way MANOVA, van Valen’s test and Box’s M test. A
Penrose's distance calculator is also implemented.
}
\references{
Manly, B.F.J., Navarro Alberto, J.A. and Gerow, K. (2024)
\emph{Multivariate Statistical Methods. A Primer}. 5th Edn. CRC Press.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ganava4/smsets}
  \item Report bugs at \url{https://github.com/ganava4/smsets/issues}
}

}
\author{
\strong{Maintainer}: Jorge Navarro Alberto \email{ganava4@gmail.com} (\href{https://orcid.org/0000-0002-5027-6908}{ORCID}) [copyright holder]

}
\keyword{internal}
