% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.BoxM.F.R
\name{print.BoxM.F}
\alias{print.BoxM.F}
\title{Prints Box'M test based on an F-statistic}
\usage{
\method{print}{BoxM.F}(x, long = FALSE, ...)
}
\arguments{
\item{x}{an object of class BoxM.F}

\item{long}{a logical variable indicating whether a long output is desired
(\code{TRUE}) or not (\code{FALSE}, the default). The long output shows the
covariance matrix for each group and the pooled covariance matrix.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Prints the results produced by BoxM.F function, with the option to display
the matrices involved in the calculations
}
\examples{
data(skulls)
resBoxM.F <- BoxM.F(skulls, Period)
# Long output
print(resBoxM.F, long = TRUE)

}
