% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.Hotelling.mat.R
\name{print.Hotelling.mat}
\alias{print.Hotelling.mat}
\title{Prints Hotelling's \eqn{T^2} test}
\usage{
\method{print}{Hotelling.mat}(x, long = FALSE, ...)
}
\arguments{
\item{x}{an object of class \code{"Hotelling.mat"}}

\item{long}{a logical variable indicating whether a long output is desired
(\code{TRUE}) or not (\code{FALSE}, the default)}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Prints the results produced by the \code{\link[smsets]{Hotelling.mat}}
function
}
\examples{
data(sparrows)
results.T2 <- Hotelling.mat(sparrows, group = Survivorship, level1 = "S")
# Long output
print(results.T2, long = TRUE)

}
