% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.VanValen.R
\name{print.VanValen}
\alias{print.VanValen}
\title{Prints van Valen's test}
\usage{
\method{print}{VanValen}(x, long = FALSE, ...)
}
\arguments{
\item{x}{an object of class VanValen}

\item{long}{a logical variable indicating whether a long output is desired
(\code{TRUE}) or not (\code{FALSE}, the default)}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Displays the results of van Valen's test produced by the \code{VanValen}
function and, optionally, the matrices involved in the calculations of van
Valen's test.
}
\examples{
data(sparrows)
res.VanValen <- VanValen(sparrows, "Survivorship", "S",
                         alternative = "less", var.equal = TRUE)
# Long output
print(res.VanValen, long = TRUE)

}
