% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smsets-package.R
\docType{package}
\name{smsets-package}
\alias{smsets}
\alias{smsets-package}
\title{smsets}
\description{
A collection of simple parameter estimation and tests for the comparison of multivariate means and variation, covered in Chapter 4 of the book Multivariate Statistical Methods. A Primer. 5th edition.
}
\details{
The goal of smsets is to produce simple multivariate statistical tests for
means and variances / covariances for one single factor with two or more
levels, including multiple two-sample t- and Levene’s tests, Hotelling’s
\eqn{T^2} test, extended two-sample Levene’s tests for multivariate data,
one-way MANOVA, van Valen’s test and Box’s M test. A Penrose's distance
calculator is also implemented.
}
\references{
Manly, B.F.J., Navarro Alberto, J.A. and Gerow, K. (2024)
\emph{Multivariate Statistical Methods. A Primer}. 5th Edn. CRC Press.
}
\author{
\strong{Maintainer}: Jorge Navarro Alberto \email{ganava4@gmail.com} (\href{https://orcid.org/0000-0002-5027-6908}{ORCID})

}
\keyword{internal}
